requirements.txt
ccxt==2.8.84
fastapi==0.95.1
uvicorn==0.21.1
prometheus-client==0.16.0
PyYAML==6.0.1
requests==2.28.2
python-dotenv==1.0.1
pydantic==1.10.14
sqlalchemy==2.0.23
aiohttp==3.9.1
python-socketio==5.10.0
psycopg2-binary==2.9.9